+++ favor/src/main/java/com/favor/favor/anniversary/Anniversary.java
@@ -1,7 +1,7 @@
 package com.favor.favor.anniversary;
 
 import com.favor.favor.common.TimeStamped;
-import com.favor.favor.member.User;
+import com.favor.favor.user.User;
 import lombok.*;
 
 import javax.persistence.*;
