+++ favor/src/main/java/com/favor/favor/friend/FriendController.java
@@ -4,7 +4,7 @@ import com.favor.favor.common.DefaultResponseDto;
 import com.favor.favor.friend.account.FriendUserRequestDto;
 import com.favor.favor.friend.noAccount.FriendRequestDto;
 import com.favor.favor.friend.noAccount.FriendUpdateRequestDto;
-import com.favor.favor.member.UserResponseDto;
+import com.favor.favor.user.UserResponseDto;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
