+++ favor/src/main/java/com/favor/favor/reminder/Reminder.java
@@ -1,7 +1,7 @@
 package com.favor.favor.reminder;
 
 import com.favor.favor.friend.Friend;
-import com.favor.favor.member.User;
+import com.favor.favor.user.User;
 import lombok.*;
 
 import javax.persistence.*;
