+++ favor/src/main/java/com/favor/favor/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.favor.favor.member;
+package com.favor.favor.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
